Product Reviews: treat rating as float
Rating is an average of a bunch of ints and thus is a float!

Otherwise, 4.9 is rounded down to 4!

+++ frontend/models/product.ts
@@ -603,7 +603,7 @@ function parseVideosJson(
 type RatingMetafield = z.infer<typeof RatingMetafieldSchema>;
 
 const stringNumberTransform = (val: string, ctx: RefinementCtx) => {
-   const parsed = parseInt(val);
+   const parsed = parseFloat(val);
    if (isNaN(parsed)) {
       ctx.addIssue({
          code: z.ZodIssueCode.custom,
